requirements.txt
autopep8==1.5.1
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-grappelli==2.14.1
idna==2.9
Pillow==7.1.1
pycodestyle==2.5.0
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stripe==2.45.0
urllib3==1.25.9
